+++ .storybook/preview.js
@@ -1,13 +1,19 @@
 import Theme from "../src/lib/Theme/Theme.svelte";
 import "$lib/style/main.css";
-import { SyntaxHighlighter } from "@storybook/components";
+import { SyntaxHighlighter } from "storybook/internal/components";
 import svelte from "./svelte-highlighting";
 
 SyntaxHighlighter.registerLanguage("svelte", svelte);
 
-/** @type { import('@storybook/svelte').Preview } */
+/** @type { import('@storybook/sveltekit').Preview } */
 const preview = {
   parameters: {
+    backgrounds: {
+      options: {
+        light: { name: "light", value: "#ffffff" },
+        dark: { name: "dark", value: "#0A4C6A" }
+      }
+    },
     controls: {
       matchers: {
         color: /(background|color)$/i,