+++ src/lib/Button/IconDownload.svelte
@@ -1,7 +1,6 @@
 <script>
   export let fill = "#FFFFFF";
   export let size = 48;
-
 </script>
 
 <svg width={size} height={size} viewBox="0 0 48 48" xmlns="http://www.w3.org/2000/svg">