+++ src/lib/ProjectCredits/ProjectCredits.stories.svelte
@@ -1,8 +1,10 @@
 <script context="module">
   import ProjectCredits from "./ProjectCredits.svelte";
+  import TextBlock from "$lib/TextBlock/TextBlock.svelte";
   import docs from "./ProjectCredits.docs.md?raw";
+  import { defineMeta } from "@storybook/addon-svelte-csf";
 
-  export const meta = {
+  const { Story } = defineMeta({
     title: "Components/ProjectCredits",
     component: ProjectCredits,
     tags: ["autodocs"],
@@ -16,13 +18,7 @@
         url: "/ProjectCredits/ProjectCredits.svelte"
       }
     }
-  };
-</script>
-
-<script>
-  import TextBlock from "$lib/TextBlock/TextBlock.svelte";
-  import { Story, Template } from "@storybook/addon-svelte-csf";
-
+  });
   const sampleData = {
     heading: "Project credits",
     items: [
@@ -47,19 +43,19 @@
   };
 </script>
 
-<Template let:args>
-  <ProjectCredits {...args}>
-    <TextBlock slot="intro"
-      ><em
-        >Vitae turpis massa sed elementum tempus. At quis risus sed vulputate odio ut enim blandit
-        volutpat. Odio ut sem nulla pharetra. Diam maecenas sed enim ut sem viverra. Nunc lobortis
-        mattis aliquam faucibus purus in massa. Vel eros donec ac odio tempor orci. Viverra tellus
-        in hac habitasse platea. Eget nunc scelerisque viverra mauris in aliquam set egestas quis.
-        Integer enim neque volutpat ac tincidunt vitae semper quis. Sociis natoque penatibus et
-        magnis dis parturient montes nascetur ridiculus.</em
-      ></TextBlock
-    >
+<Story name="Default" asChild>
+  <ProjectCredits {...sampleData}>
+    {#snippet intro()}
+      <TextBlock
+        ><em
+          >Vitae turpis massa sed elementum tempus. At quis risus sed vulputate odio ut enim blandit
+          volutpat. Odio ut sem nulla pharetra. Diam maecenas sed enim ut sem viverra. Nunc lobortis
+          mattis aliquam faucibus purus in massa. Vel eros donec ac odio tempor orci. Viverra tellus
+          in hac habitasse platea. Eget nunc scelerisque viverra mauris in aliquam set egestas quis.
+          Integer enim neque volutpat ac tincidunt vitae semper quis. Sociis natoque penatibus et
+          magnis dis parturient montes nascetur ridiculus.</em
+        ></TextBlock
+      >
+    {/snippet}
   </ProjectCredits>
-</Template>
-
-<Story name="Default" args={{ ...sampleData }} />
+</Story>